Fixed tricky popover bug
