+++ assets/src/containers/AssignmentPlanningV2.js
@@ -94,6 +94,7 @@ function AssignmentPlanningV2 (props) {
       ...assignments.slice(key + 1)
     ])
   }
+
   const handleClearGoalGrades = () => {
     setAssignments(
       assignments.map(a => {
